Major rewrite! Ready to try hosting.
This rewrites the server with Typescript and Prisma instead of direct sql queries. There are now sessions and users, https, nginx, logging. I did the same thing I did on Overseer and Context.Reviews for everything.

The dates should be more reliable. Between the server and client every date should be serialized to a ISO 8601 date with no time, so timezone differences between the server and client shouldn't matter.

The todo routes follow rest conventions a lot better.