Added the ability for users to update their name/email/password.
To do this I needed to serialize the user info when the page loads, which required the introduction of the `serialize-javascript` package and additional locals on the app's template.

I also added `page` and some components to handle routing, as there's now a `/settings` route to show the profile/password change page. Routing is handled by setting a route slug to a `activeRoute` store. Depending on that route the main contents on the page will mount a different component (Settings or Todos is all at the moment).

I moved some of the auth routes to the user router because they fit closer to some of the new ones for updating the profile and stuff. I had to move some utils to a different folder because circular references broke imports!