Error handling for adding todos.
Previously when you'd add a todo the todolist component would just
delete the todo text you typed which is fine unless it errors. The
validation for todos has a character limit that's somewhat small so if
someone went through the trouble of typing out a bunch it kinda sucks to
have it all deleted, and without warning because it wouldn't notify you
that it failed. Now it won't clear the field unless it was successful
and if it wasn't it'll show an error toast saying your todo is too long.