Todo validation changes.
This ups the character limit for a todo from 100 to 300. That's probably
unrealistically long, but I'm going to leave that up to the user. Also
made the "what list" part of the todo validation schema more strict so
it actually only allows the lists Konshuu provides.

I also removed a bunch of awaits typescript complained were unnecessary.