Mobile display improvements
The mobile display wasn't the greatest. This fixes some scrolling
issues by allowing days to scroll individually, and putting the new todo
form at the top of the list so it won't be hidden by the browser's
floating address bar on mobile firefox where it's by default on the
bottom of the page.

Also made the lists be a max of no wider than the viewport on mobile, so
it shouldn't allow lists with long items to get too wide.