Added singular phrasing for the reschedule modal
The reschedule modal was built with the intention of rescheduling multiple todos (for an entire day at a time) but it is just as likely to be used for a single todo. This change makes the phrasing for pluralizing the phrasing on the reschedule modal to react the the nummber of todos it is responsible for rescheduling.