Switch to a "+" button on mobile instead of an ever present text input.
The new todo text input was kind of wonky on mobile, Firefox liked to zoom in on it for some reason. The only way I could get it to mostly not do that was to make the font size really big, but now it's unnaturally big and it looks weird and takes up a lot of space. Also when the keyboard opens up it can still cover the list or the text input. Switching this to a native prompt modal should hopefully help. The prompt modal is triggered by clicking the new "+" button at the top of a todo list, next to the reschedule button.

Also added a bit more error handling, this will show a generic error for non-400 errors.