Rewrote communications with websockets
The client/server communication was really inefficent. The page would poll every 10 seconds for the entire week's todos, and again after any operation (completing/editing/deleting/rescheduling a todo) so it would request the week's todos very often and most of the time it wouldn't have changed. This switches the whole site over to websockets. When a websocket connects they are automatically put into a socket.io room for their user, any operations that mutate their list will be broadcast to that room so the user will get updates for any change. Additionally every operation now sends a fairly minimal amount of data instead of the entire list. Depending on the operation the store will be updated accordingly. This allows no additional queries to hit the database for the whole week's todos for normal todo usage.

The only time a full week query is run is when the current week changes, the offset changes (previous/next week is viewed) or when the websocket reconnects.