Added shortcuts to navigate lists/weeks.
When on a computer you can use Ctrl+Alt+Arrows to switch which new todo input/list you're focused on. Trying to go beyond the ends of a week will go to to the next/previous week (if pressing right on a Saturday, it'll go to the following Sunday in the next week).

On mobile there are "Next/Previous Week" buttons that show up below the list on weekends that will allow you to quickly go to the adjacent day that weekend while crossing the week boundary.