Pass socket communications through Onaji
Onaji lets data be serialized and parsed using the original class types on both ends. I'm using this for Date and CalendarDate objects so they don't need to be manually serialized/deserialized imperatively all over the place. This also saves on needing two typings (serialized and non-serialized versions) of todos, much cleaner.

I also added a "Next Week" reschedule shortcut which is just the same day next week, and all of the options that take target an explicit day of the week now do it with respect to the date that todo is from, so if you're looking three weeks in the future and reschedule a todo to "Saturday" it'll be that week's Saturday, not Saturday of the current week.