Allow dragging todos to a different list
This doesn't let you bulk move to a different list, but at least if you drop a todo on the specified list it'll go to the list you drop it on instead of the adjacent one. Before it'd reschedule to that day, but it'd stay in the list it was originally.

This also adds a blue border when dragging on the list which looks a little nicer.