Add weeklies
This is a separate kind of todo that needs to be completed a set number of times any time during the week. They repeat every week and can be individually edited per week to deviate from the normal goal.

Eventually I want to use the same strategy here with recurring todos. There are two new database models, a Weekly and WeeklyProgress. Weekly stores the name/goal, and WeeklyProgress is the number of times a weekly has been completed in a given week. Until a weekly has been completed once the user sees placeholders for progress for each week, and when they complete a weekly (or edit that week's goal) it will be created.