Only sent weeklies that are currently active or have progress for the week
This optimizes responses for users who have a lot of deleted weeklies. We were previously querying for all weeklies a user has ever made and only showing ones that are active, but if they've deleted a weekly and don't have progress for the week they're viewing it's not useful to see what they had before. This will filter for non-deleted weeklies and merge in the list of included weeklies when querying for weekly progress. That way we're neither querying for or sending weeklies that aren't relevant.