Added countdown anytime time
This new type lets you specify a date and time to count down to. It will show the difference in text (e.g. "4 days 2 hours 44 minutes"), then once the timer is reached the text changes color and a delete button shows up (so you can clear the timer without having to go into the settings).

Also changed the ellipsis icon for settings to a cog. Now that a green color is used when the timer expires I wanted less colors that have no meaning so counters just have a gray background instead of pink.