Allow 0 goal weeklies, allow updating weeklies with no progress
Sometimes you might want to skip a weekly so it should be valid to set a 0 goal for a week, lowered the minimum allowed goal to 0.

I had been extra defensive with IDs in a previous change, but forgot that weeklies should be able to upsert in which case an id will be null at first. This lets reverts that change.