Match weekly progress by week
Weekly progress updates were only checking that the weeklyId matched, but if the user had two tabs open on different weeks progress for one weekly would be broadcast to both and overwrite the wrong week's progress on the tab that wasn't looking at that week. Now it will check to ensure the week and weeklyId match before applying an update.