Added a countup anytime
This is very similar to the countdown (and uses the same `countdownEnd` database column) but it doesn't have an "end". It counts the time since a date, and the settings let you reset the date to now. This is good for tracking streaks.

Because there is no end to a countup, it doesn't have an "expired" state, and doesn't show a delete button outside of the settings panel like expired countdowns have.