Track the time an Anytime was last updated
For edit/increment/decrement the new `updatedAt` field will be set to the current time.

When increment/decrement messages for counters are received by the browser it will set updatedAt to a new Date. This might be slightly off from the actual updatedAt, but those are atomic updates and to avoid data coming back out of order if the user clicks quickly it computes the count instead of sending an updated count. As such we don't have the real updatedAt so just faking it client side should be fine.