Recurring todos
This is a new feature where you can set up todos with various schedules. The options are similar to what Habitica has for dailies, which is what I used for recurring todos previously.

You make a "recurring todo", and every list checks if any todos match that list/date, for anything that matches it will show a todo with just a checkbox. The existence of a separate "recurring todo completion" for that recurring todo id and date means it has been completed, and checking/unchecking the box deletes and creates a row in that completion table.

The UI for all of the recurring todo creation stuff is in a new "sidebar" that lets you switch from weeklies to that.

I had to upgrade typescript/eslint stuff because it didn't realize that arrays have an `at` method now.