Don't just hope the sidebar is off screen when it's closed
This adds `display: none` to the sidebar once it's fully closed, so it won't impact the layout of the page. While it was uninteractable and invisible, if the sidebar was taller than 20rem (the negative margin it gets off screen when closed) it would still impact the position of anytimes below it.