Less prominent checked todos, less obnoxious recurring todo container
I copied the checkbox component from sheodox-ui and made a TodoCheckbox based on it, which uses muted colors when checked instead of a bright blue checkbox which was counterproductive to the whole getting-out-of-the-way idea I was going for once you finish something. This also lets me get rid of all the hacky `:global` overrides I had been using to get the styling I had previously.

I also removed the border on the recurring todo container in a todo list, so it's still obvious it's recurring todos, but it doesn't feel look like a focus outline, and it's not obnoxiously prominent.