Only show the reschedule drop indicator when actually dragging a todo
My computer seems to be sending phantom drag events when I switch focus between windows, so randomly a todo list is getting the dragging over list indicator. This makes it only show that if the user is actually dragging a todo, so weird phantom drag events or when the user drags a file over the Konshuu window or something won't trigger the indicator anymore.