Migrate to using sheodox-ui's sidebar for Anytimes
This is much better integrated with the UI in general and looks nicer.

I also made the '<- Show All Anytimes' link stay there all the time and look more like another tag, except for the different icon and buttons. It also mirrors the "All Anytimes" button that shows in the toolbar.