Don't show the new anytime todo form automatically
If an anytime todo list has no todos, it now shows a message and has a button to make a new todo. This allows us to make the name field auto focusing without it trying to focus automatically when the user just views a page with an empty todo list, it now requires them to take an action to attempt adding todos before it autofocuses.