Focus '+' button when cancelling the anytime todo form, sheodox-ui update
This allows you to easily add another todo if you closed the form but haven't clicked elsewhere, just hit spacebar or enter again to reopen the form, just like the week todos.

Updating sheodox-ui so if you load the page and an expired countdown shows right away, it can be dismissed without having to delete the countdown.